products-service/requirements.txt
alembic==1.18.4
asyncpg==0.31.0
fastapi==0.135.1
greenlet==3.3.2
httpx==0.28.1
pydantic==2.12.5
python-dotenv==1.2.2
sqlalchemy==2.0.48
uvicorn==0.42.0
